+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog of IARE
 
+## v0.2.24.11
+- Add TLD (Top Level Domains) chart
+- placed charts in 2 columns
+
 ## v0.2.24.10
 - Added version display for Production environment
 - Consolidate Chart Legend code